@uppy/audio: refactor to ESM
npm package is still CommonJS, the plan is to switch that in the
next semver-major.
