+++ packages/@uppy/audio/package.json
@@ -16,6 +16,7 @@
     "record",
     "mediarecorder"
   ],
+  "type": "module",
   "homepage": "https://uppy.io",
   "bugs": {
     "url": "https://github.com/transloadit/uppy/issues"
