+++ packages/@uppy/audio/src/formatSeconds.js
@@ -5,7 +5,7 @@
  * @returns {string} the formatted seconds (e.g. '1:23' for 1 minute and 23 seconds)
  *
  */
-module.exports = function formatSeconds (seconds) {
+export default function formatSeconds (seconds) {
   return `${Math.floor(
     seconds / 60,
   )}:${String(seconds % 60).padStart(2, 0)}`
