+++ packages/@uppy/audio/src/formatSeconds.test.js
@@ -1,4 +1,4 @@
-const formatSeconds = require('./formatSeconds')
+import formatSeconds from './formatSeconds.js'
 
 describe('formatSeconds', () => {
   it('should return a value of \'0:43\' when an argument of 43 seconds is supplied', () => {
