+++ packages/@uppy/audio/src/supportsMediaRecorder.js
@@ -1,4 +1,4 @@
-module.exports = function supportsMediaRecorder () {
+export default function supportsMediaRecorder () {
   /* eslint-disable compat/compat */
   return typeof MediaRecorder === 'function'
     && typeof MediaRecorder.prototype?.start === 'function'
